changed file names and used card components on both creatures and gods
